Change get_dataroot to return None when none exists.
